+++ index.js
@@ -6,490 +6,17 @@
  *
  */
 
-/* eslint-disable */
 // importing node package
-const https = require("https");
-const http = require("http");
-
-function DNSSEC(host, callback) {
-    // removing url scheme is present
-    host = /^http(s)?:\/\//.test(host) ? host.split("://")[1] : host;
-
-    // forming url
-    let url = `https://viewdns.info/dnssec/?domain=${host}`;
-
-    // firing get request
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response chunk encoding
-
-            let body = ""; // declaring variable to
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // the chunk collection is over now
-                let data = {
-                    public: [],
-                    dnssecSignatures: []
-                }
-
-                try {
-                    // performing the magic
-                    let part1 = body.split("Below are the DNSSEC public keys for this domain:")[1]
-                    let part2 = body.split("Below are the DNSSEC signatures for this domain:")[1]
-                    if (part1 && part2) {
-                        part1
-                            .split("<table width=\"900\" border=\"1\">")[1]
-                            .split("</table>")[0]
-                            .split("<tr><td>")
-                            .splice(2)
-                            .forEach(txt => {
-                                let _ = txt.split("</td><td>");
-                                data.public.push({
-                                    keyType: _[0],
-                                    algorithm: _[1],
-                                    key: _[2].split("<br></td></tr>")[0].replace(/<br>/gm, "\n")
-                                })
-
-                            });
-                        part2
-                            .split("<table width=\"900\" border=\"1\">")[1]
-                            .split("</table>")[0]
-                            .split("<tr><td>")
-                            .splice(2)
-                            .forEach(txt => {
-                                let _ = txt.split("</td><td>");
-                                data.dnssecSignatures.push({
-                                    recordType: _[0],
-                                    algorithm: _[1],
-                                    signedBy: _[2],
-                                    signature: _[3].split("<br></td></tr>")[0].replace(/<br>/gm, "\n")
-                                })
-
-                            });
-                        return callback(null, data)
-                    } else {
-                        return callback(null, "This domain DOES NOT have DNSSEC enabled.");
-                    }
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function ReverseIPLookup(host, callback) {
-    // removing url scheme is present
-    host = /^http(s)?:\/\//.test(host) ? host.split("://")[1] : host;
-
-    // forming url
-    let url = `https://viewdns.info/reverseip/?host=${host}&t=1`;
-
-    // firing get request
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response chunk encoding
-
-            let body = ""; // declaring variable to
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // the chunk collection is over now
-                let data = [];
-
-                try {
-                    // performing the magic
-                    body
-                        .split(/<table (.*)="1">/)[2]
-                        .split(/<\/table>/)[0]
-                        .split(" <td>")
-                        .splice(1)
-                        .forEach(domain => {
-                            data.push({
-                                domain: domain.split("<")[0],
-                                lastResolved: domain.split("\">")[1].split("<")[0]
-                            }); // splitting ad getting required data only
-                        });
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-
-                return callback(null, data);
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function IP2Location(ip, callback) {
-    if (ip == "") return callback(new Error("No IP passed"), null);
-
-    http
-        .get(`http://ip-api.com/json/${ip}`, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                return callback(null, JSON.parse(body));
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function Down(host, callback) {
-    let url = `https://viewdns.info/ismysitedown/?domain=${host}`;
-
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                return callback(null, /not accessible/.test(body));
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function ASN(asn, callback) {
-    let url = `https://viewdns.info/asnlookup/?asn=${asn}`;
-
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                try {
-                    // performing the magic
-                    let d = {};
-                    body
-                        .split("==============<br><br>")[1]
-                        .split("<br><br><br></td>")[0]
-                        .split("<br>")
-                        .filter(x => x != "")
-                        .forEach(el => {
-                            const [a, b] = el.split(": ");
-                            d[a.toLowerCase()] = b.trimLeft().toLowerCase();
-                        });
-
-                    return callback(null, d);
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function MXLookup(mailServer, callback) {
-    let url = `https://viewdns.info/reversemx/?mx=${mailServer}`;
-
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                try {
-                    // performing the magic
-                    let d = [];
-                    body
-                        .split("<table border=\"1\">")[1]
-                        .split("</table>")[0]
-                        .split("<tr><td>")
-                        .forEach(domain => {
-                            const _domain = domain.split("</td></tr>")[0];
-                            d.push(_domain);
-                        });
-
-                    return callback(null, {
-                        domains: d.splice(2)
-                    });
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function CFTest(host, callback) {
-    let url = `https://viewdns.info/chinesefirewall/?domain=${host}`;
-
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                try {
-                    // performing the magic
-                    let d = [];
-                    body
-                        .split("<table border=\"1\">")[1]
-                        .split("</table>")[0]
-                        .split("<tr><td>")
-                        .splice(1)
-                        .forEach(entry => {
-                            const _entry = /\/images\/ok/.test(entry) ?
-                                entry
-                                .replace(
-                                    "<img src=\"/images/ok.GIF\" height=\"20\" alt=\"ok\">",
-                                    false
-                                )
-                                .replace("<center>", "")
-                                .replace("</center></td></tr>", "")
-                                .replace("<br>", "")
-                                .split("</td><td>") :
-                                entry
-                                .replace(
-                                    "<img src=\"/images/error.GIF\" height=\"20\" alt=\"fail\">",
-                                    true
-                                )
-                                .replace("<center>", "")
-                                .replace("</center></td></tr>", "")
-                                .replace("<br>", "")
-                                .split("</td><td>");
-                            d.push({
-                                location: _entry[0].trim(),
-                                lookupResult: _entry[1].trim(),
-                                isBlocked: _entry[2] == "false" ? false : true
-                            });
-                        });
-
-                    return callback(null, d);
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function PortScan(host, callback) {
-    let url = `https://viewdns.info/portscan/?host=${host}`;
-
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                try {
-                    // performing the magic
-                    let d = [];
-                    body
-                        .split("<table border=\"1\">")[1]
-                        .split("</table>")[0]
-                        .split("<tr><td>")
-                        .splice(1)
-                        .forEach(entry => {
-                            const _entry = /\/images\/ok/.test(entry) ?
-                                entry
-                                .replace(
-                                    "<img src=\"/images/ok.GIF\" height=\"20\" alt=\"open\">",
-                                    true
-                                )
-                                .replace("<center>", "")
-                                .replace("</center></td></tr>", "")
-                                .replace("<br>", "")
-                                .split("</td><td>") :
-                                entry
-                                .replace(
-                                    "<img src=\"/images/error.GIF\" height=\"20\" alt=\"closed\">",
-                                    false
-                                )
-                                .replace("<center>", "")
-                                .replace("</center></td></tr>", "")
-                                .replace("<br>", "")
-                                .split("</td><td>");
-                            d.push({
-                                port: _entry[0].trim(),
-                                service: _entry[1].trim(),
-                                isOpen: _entry[2] == "false" ? false : true
-                            });
-                        });
-
-                    return callback(null, d);
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
-function IPHistory(host, callback) {
-    let url = `https://viewdns.info/iphistory/?domain=${host}`;
-
-    https
-        .get(url, res => {
-            res.setEncoding("utf8"); // setting response text encoding
-
-            let body = ""; // inititializing body
-
-            res.on("data", chunk => {
-                // collecting chunks
-                body += chunk;
-            });
-
-            res.on("error", e => {
-                // handling error while reading response
-                return callback(e, null);
-            });
-
-            res.on("end", () => {
-                // validate and return data on end of response read
-                try {
-                    // performing the magic
-                    let d = [];
-                    body
-                        .split("<table border=\"1\">")[1]
-                        .split("</table>")[0]
-                        .split("<tr><td>")
-                        .splice(1)
-                        .forEach(entry => {
-                            const _entry = entry.split("</td><td>");
-                            d.push({
-                                ip: _entry[0],
-                                location: _entry[1],
-                                ipOwner: _entry[2].split("\r\n")[0].replace("&#44;", ","),
-                                lastSeen: _entry[2].split("\">")[1].split("</td></tr>")[0]
-                            });
-                        });
-
-                    return callback(null, d);
-                } catch (e) {
-                    // handling the error
-                    return callback(e, null);
-                }
-            });
-        })
-        .on("error", e => {
-            // return error callback if error related to firing requests
-            return callback(e, null);
-        });
-}
-
+const DNSReport = require("./dnsreport");
+const DNSSEC = require("./dnssec");
+const IP2Location = require("./ip2loc");
+const ReverseIPLookup = require("./reverseip");
+const Down = require("./ismyhostdown");
+const ASN = require("./asn");
+const MXLookup = require("./mx");
+const CFTest = require("./cftest");
+const PortScan = require("./portscan");
+const IPHistory = require("./iphistory");
 // exporting modules
 module.exports = {
     /**
@@ -561,10 +88,18 @@ module.exports = {
     ipHistory: IPHistory,
     /**
      * DNSSEC Test
-     * 
+     *
      * @description Test if any domain name is configured for DNSSEC (Domain Name System Security Extensions).
      * @param {string} host
      * @param {callback} callback
      */
-    dnssec: DNSSEC
+    dnssec: DNSSEC,
+    /**
+     * DNS Report
+     *
+     * @description View a complete report on the DNS settings for your domain. This tool is designed to assist webmasters and system administrators diagnose DNS related issues. A number of tests are run on your DNS settings with results displayed in an easy to understand manner.
+     * @param {string} host
+     * @param {callback} callback
+     */
+    dnsreport: DNSReport
 };
