+++ test.js
@@ -65,4 +65,11 @@ describe("Test Cases", () => {
             done();
         });
     }).timeout(15000);
+
+    it("DNS Report", done => {
+        info.dnsreport("cybitrock.com", err => {
+            expect(err).to.be.equal(null);
+            done();
+        });
+    }).timeout(15000);
 });
